moves operator API into separate package it should be easier to use it as go lib
