+++ api/v1beta1/vmagent_types.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/VictoriaMetrics/operator/controllers/factory/crd"
 	appsv1 "k8s.io/api/apps/v1"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -560,7 +559,7 @@ func (cr *VMAgent) AsURL() string {
 
 // AsCRDOwner implements interface
 func (cr *VMAgent) AsCRDOwner() []metav1.OwnerReference {
-	return crd.GetCRDAsOwner(crd.VMAgent)
+	return GetCRDAsOwner(Agent)
 }
 
 func init() {
