+++ api/v1beta1/vmalert_types.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/VictoriaMetrics/operator/controllers/factory/crd"
 	appsv1 "k8s.io/api/apps/v1"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -465,7 +464,7 @@ func (cr *VMAlert) AsURL() string {
 
 // AsCRDOwner implements interface
 func (cr *VMAlert) AsCRDOwner() []metav1.OwnerReference {
-	return crd.GetCRDAsOwner(crd.VMAlert)
+	return GetCRDAsOwner(Alert)
 }
 
 func (cr *VMAlert) GetNotifierSelectors() []*DiscoverySelector {
