+++ api/v1beta1/vmalertmanager_types.go
@@ -4,7 +4,6 @@ import (
 	"fmt"
 	"strings"
 
-	"github.com/VictoriaMetrics/operator/controllers/factory/crd"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
@@ -375,7 +374,7 @@ func (cr *VMAlertmanager) MetricPath() string {
 
 // AsCRDOwner implements interface
 func (cr *VMAlertmanager) AsCRDOwner() []metav1.OwnerReference {
-	return crd.GetCRDAsOwner(crd.VMAlertManager)
+	return GetCRDAsOwner(AlertManager)
 }
 
 // AsNotifiers converts VMAlertmanager into VMAlertNotifierSpec
