+++ api/v1beta1/vmcluster_types.go
@@ -5,7 +5,6 @@ import (
 	"path"
 	"strings"
 
-	"github.com/VictoriaMetrics/operator/controllers/factory/crd"
 	appsv1 "k8s.io/api/apps/v1"
 	v1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
@@ -941,7 +940,7 @@ func (cr *VMCluster) VMStorageURL() string {
 
 // AsCRDOwner implements interface
 func (cr *VMCluster) AsCRDOwner() []metav1.OwnerReference {
-	return crd.GetCRDAsOwner(crd.VMCluster)
+	return GetCRDAsOwner(Cluster)
 }
 
 func (cr VMCluster) GetNSName() string {
